icu 58.2 (release), original sources
git-svn-id: svn://tug.org/texlive/trunk/Build/source@43491 c570f23f-e606-0410-a88d-b1316a301751

+++ libs/icu/icu-src/source/test/cintltst/sorttest.c
@@ -1,3 +1,5 @@
+// Copyright (C) 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
 /*
 *******************************************************************************
 *
@@ -51,7 +53,7 @@ SortTest() {
 
     /* for medium, add bits that will not be compared, to test stability */
     for(i=0; i<UPRV_LENGTHOF(medium); ++i) {
-        medium[i]=(medium[i]<<4)|i;
+        medium[i]=(int32_t)((uint32_t)medium[i]<<4) | i;
     }
 
     /* sort medium array (stable) */
